Add TanHNeuron

+++ src/main/java/com/vgalloy/neuron/neuron/Neuron.java
@@ -1,9 +1,6 @@
 package com.vgalloy.neuron.neuron;
 
 import java.util.List;
-import java.util.function.Function;
-
-import com.vgalloy.neuron.constant.Constant;
 
 /**
  * Created by Vincent Galloy on 01/04/17.
@@ -14,5 +11,5 @@ public interface Neuron {
 
     boolean apply(final List<Boolean> input);
 
-    List<Long> train(List<Boolean> input, boolean expected);
+    List<Double> train(List<Boolean> input, boolean expected);
 }
